js build unit tests done

+++ src/utils/browserify-helper.js
@@ -1,13 +1,24 @@
 var config = require('./get-config')();
 var GetPath = require('./get-path');
 var shelljs = require('shelljs');
+var path = require('path');
+var PathExists = require('./path-exists');
+var glob = require('glob');
 
 // dist, src, watch, standalone
 var browserifyHelper = function(options) {
   ['.js', '.js.map'].forEach(function(ext) {
-    shelljs.rm('-rf', options.dist + ext);
+    var file = options.dist + ext;
+    if (PathExists(file)) {
+      shelljs.rm('-rf', options.dist + ext);
+    }
   });
+  var oldSilent = shelljs.config.silent;
   shelljs.mkdir('-p', path.dirname(options.dist));
+
+  shelljs.config.silent = true;
+  shelljs.pushd(path.join(__dirname, '..', '..'));
+  shelljs.config.silent = oldSilent;
 
   var babelPreset = config.ie8 ? GetPath('babel-preset-ie8.config.js') : GetPath('babel-preset.config.js');
   var files = [options.src];
@@ -19,16 +30,30 @@ var browserifyHelper = function(options) {
   var browserify = (options.watch ? GetPath('watchify') + ' -v' : GetPath('browserify'))
     + ' --debug'
     + ' -t rollupify'
-    + ' -t [ babelify --preset ' + babelPreset + ' ]'
+    + ' -t [ babelify --presets ' + babelPreset + ' ]'
     + ' -g browserify-shim'
     + ' -g browserify-versionify'
     + ' -p bundle-collapser/plugin'
-    + ' -p [ mapstraction ' + options.dist + '.js.map ]'
+    + ' -p mapstraction'
     + (config.standalone ? ' -s ' + config.name : '')
     + ' -o ' + options.dist + '.js'
-    + files.join(' ');
+    + ' ' + files.join(' ');
 
-  shelljs.exec(browserify);
+  var retval = shelljs
+    .exec(browserify, {silent: true});
+
+  // remove rollup external deps errors...
+  retval.stderr = retval.stderr || '';
+  retval.stderr = retval.split(/^Treating '\w+' as external dependency\n$/).join('');
+
+  if (retval.stderr) {
+    process.stderr.write(retval.stderr);
+  }
+
+  shelljs.config.silent = true;
+  shelljs.popd();
+  shelljs.config.silent = oldSilent;
+  return retval;
 };
 
 module.exports = browserifyHelper;
